+++ src/onegov.file/HISTORY.rst
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+- Changes frontend cache busting to use ORM events.
+  [href]
+
 0.7.0 (2018-08-06)
 ~~~~~~~~~~~~~~~~~~~